Fixed regression introduced in r18356 (CVE-2011-4612): client duration time is now correctly logged. PRIu64 MUST NOT be used with log_write_direct() as depending on platform PRIu64 may be using something not supported by __vsnprintf() of log/log.c. close #2081, see r18356
svn path=/icecast/trunk/icecast/; revision=19287

+++ src/logging.c
@@ -152,7 +152,7 @@ void logging_access(client_t *client)
         user_agent = "-";
 
     log_write_direct (accesslog,
-            "%s - %H [%s] \"%H %H %H/%H\" %d %" PRIu64 " \"% H\" \"% H\" %lu",
+            "%s - %H [%s] \"%H %H %H/%H\" %d %llu \"% H\" \"% H\" %llu",
             client->con->ip,
             username,
             datebuf,
@@ -161,10 +161,10 @@ void logging_access(client_t *client)
             httpp_getvar (client->parser, HTTPP_VAR_PROTOCOL),
             httpp_getvar (client->parser, HTTPP_VAR_VERSION),
             client->respcode,
-            client->con->sent_bytes,
+            (long long unsigned int)client->con->sent_bytes,
             referrer,
             user_agent,
-            (unsigned long)stayed);
+            (long long unsigned int)stayed);
 }
 /* This function will provide a log of metadata for each
    mountpoint.  The metadata *must* be in UTF-8, and thus
